optimizare rulare correction_valuation

+++ l10n_ro_stock_account_check/models/stock_valuation_layer.py
@@ -24,7 +24,7 @@ class StockValuationLayer(models.Model):
                     continue
 
 
-    def correction_valuation(self):
+    def correction_valuation(self, unlink_account_move=True):
         for svl in self:
             if not svl.stock_move_id:
                 continue
@@ -39,10 +39,18 @@ class StockValuationLayer(models.Model):
                 svsl_vals = product._prepare_out_svl_vals(abs(svl.quantity), svl.company_id)
                 svl.write({"unit_cost": svsl_vals["unit_cost"], "value": svsl_vals["unit_cost"] * svl.quantity})
 
-            svl.account_move_id.write({"state": "draft"})
-            name = svl.account_move_id.name
-            svl.account_move_id.with_context(force_delete=True).unlink()
-            stock_move = svl.stock_move_id.with_context(force_period_date=svl.stock_move_id.date)
-            stock_move._account_entry_move(svl.quantity, svl.description, svl.id, svl.value)
-            if name and svl.account_move_id:
-                svl.account_move_id.write({"name": name})
+            if unlink_account_move:
+                svl.account_move_id.write({"state": "draft"})
+                name = svl.account_move_id.name
+                svl.account_move_id.with_context(force_delete=True).unlink()
+                stock_move = svl.stock_move_id.with_context(force_period_date=svl.stock_move_id.date)
+                stock_move._account_entry_move(svl.quantity, svl.description, svl.id, svl.value)
+                if name and svl.account_move_id:
+                    svl.account_move_id.write({"name": name})
+            else:
+                value = abs(svl.value)
+                for line in svl.account_move_id.line_ids:
+                    if line.debit:
+                        line.with_context(check_move_validity=False).write({'debit':value})
+                    else:
+                        line.with_context(check_move_validity=False).write({'credit':value})
